Workaround possible scheduling race error
On SMP systems, there is a scheduling race that must be dealt with. In extremely
unlucky circumstances, it’s possible that the current thread disables the WP bit,
is scheduled out at a precise moment, is re-scheduled onto a CPU that still has
the WP bit enabled, and faults when attempting to perform a write to read-only
memory.

http://vulnfactory.org/blog/2011/08/12/wp-safe-or-not/